# docs/swagger.yaml
basePath: /
definitions:
  frameworkdto.CreatedResponseDTO:
    description: Created response structure
    properties:
      data: {}
      message:
        example: Resource created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  frameworkdto.ErrorDetailDTO:
    description: Error detail structure
    properties:
      code:
        example: ERR_001
        type: string
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  frameworkdto.ErrorResponseDTO:
    description: Error API response structure
    properties:
      error:
        $ref: '#/definitions/frameworkdto.ErrorDetailDTO'
      success:
        example: false
        type: boolean
    type: object
...
  frameworkdto.TenantRegistrationDTO:
    properties:
      address:
        type: string
      email:
        type: string
      licence_type_id:
        type: integer
      name:
        type: string
      phone:
        type: string
      user:
        properties:
          email:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          password:
            type: string
        type: object
    type: object
...
  frameworkdto.UserRegistrationDTO:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@geekible.com
    name: API Support
  description: |-
    A comprehensive service framework API with multi-tenant support, user management, authentication, and authorization.

    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Error Responses
    All error responses follow a standard format with success, error code, and message fields.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Service Framework API
  version: "1.0"
paths:
...
  /registration/tenant:
    post:
      consumes:
      - application/json
      description: Register a new tenant organization with an admin user
      parameters:
      - description: Tenant registration details
        in: body
        name: tenantDTO
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.TenantRegistrationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Tenant registered successfully
          schema:
            $ref: '#/definitions/frameworkdto.CreatedResponseDTO'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "409":
          description: Tenant already exists
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      summary: Register a new tenant
      tags:
      - Registration
  /registration/user:
    post:
      consumes:
      - application/json
      description: Add a new user to the authenticated tenant (requires authentication)
      parameters:
      - description: User registration details
        in: body
        name: userDTO
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.UserRegistrationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User added successfully
          schema:
            $ref: '#/definitions/frameworkdto.CreatedResponseDTO'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Add a new user to tenant
      tags:
      - Registration
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"